# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Essential HICS API
  description: |
    Hospital Incident Command System — Phase 1 REST API.
    Covers authentication, RBAC, organizational structure, and audit logging.
  version: 1.0.0
  contact:
    name: HICS Support
    email: support@apexhealth.example
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.hics.yourhospital.org
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        displayName: { type: string }
        jobTitle: { type: string }
        employeeId: { type: string }
        phoneMobile: { type: string }
        phoneWork: { type: string }
        pagerNumber: { type: string }
        authProvider: { type: string, enum: [LOCAL, AZURE_AD, OKTA] }
        isActive: { type: boolean }
        isLocked: { type: boolean }
        lastLoginAt: { type: string, format: date-time }
        passwordChangedAt: { type: string, format: date-time }
        mustChangePassword: { type: boolean }
        mfaEnabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        userFacilityRoles:
          type: array
          items: { $ref: '#/components/schemas/UserFacilityRole' }
    UserFacilityRole:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        facilityId: { type: string, format: uuid }
        hicsRole: { $ref: '#/components/schemas/HicsRole' }
        isPrimaryFacility: { type: boolean }
        assignedAt: { type: string, format: date-time }
        facility:
          type: object
          properties:
            name: { type: string }
            shortName: { type: string }
    HicsRole:
      type: string
      enum:
        - SYSTEM_ADMIN
        - SYSTEM_VIEWER
        - FACILITY_ADMIN
        - INCIDENT_COMMANDER
        - DEPUTY_INCIDENT_COMMANDER
        - PUBLIC_INFORMATION_OFFICER
        - SAFETY_OFFICER
        - LIAISON_OFFICER
        - MEDICAL_TECHNICAL_SPECIALIST
        - OPERATIONS_SECTION_CHIEF
        - PLANNING_SECTION_CHIEF
        - LOGISTICS_SECTION_CHIEF
        - FINANCE_ADMIN_SECTION_CHIEF
        - MEDICAL_CARE_BRANCH_DIRECTOR
        - INFRASTRUCTURE_BRANCH_DIRECTOR
        - SECURITY_BRANCH_DIRECTOR
        - BUSINESS_CONTINUITY_BRANCH_DIRECTOR
        - HAZMAT_BRANCH_DIRECTOR
        - PATIENT_FAMILY_INFORMATION_BRANCH_DIRECTOR
        - RESOURCES_UNIT_LEADER
        - SITUATION_UNIT_LEADER
        - DOCUMENTATION_UNIT_LEADER
        - DEMOBILIZATION_UNIT_LEADER
        - SERVICE_BRANCH_DIRECTOR
        - SUPPORT_BRANCH_DIRECTOR
        - COMMUNICATIONS_UNIT_LEADER
        - IT_SYSTEMS_UNIT_LEADER
        - STAFF_FOOD_WATER_UNIT_LEADER
        - TRANSPORTATION_UNIT_LEADER
        - FACILITIES_UNIT_LEADER
        - SUPPLY_UNIT_LEADER
        - LABOR_POOL_CREDENTIALS_UNIT_LEADER
        - TIME_UNIT_LEADER
        - PROCUREMENT_UNIT_LEADER
        - COMPENSATION_CLAIMS_UNIT_LEADER
        - COST_UNIT_LEADER
        - RESPONDER
        - READ_ONLY_OBSERVER
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
security:
  - bearerAuth: []
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List users (paginated, filterable)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 25, maximum: 100 }
        - in: query
          name: facilityId
          schema: { type: string, format: uuid }
        - in: query
          name: role
          schema: { $ref: '#/components/schemas/HicsRole' }
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive, locked] }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404': { description: User not found }

    put:
      tags: [Users]
      summary: Update user profile
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string }
                lastName: { type: string }
                displayName: { type: string }
                jobTitle: { type: string }
                phoneMobile: { type: string }
                phoneWork: { type: string }
                pagerNumber: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /api/users/{id}/deactivate:
    patch:
      tags: [Users]
      summary: Deactivate a user (revokes all sessions)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: User deactivated }
  /api/users/{id}/roles:
    post:
      tags: [Users]
      summary: Assign a HICS role at a facility
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [facilityId, hicsRole]
              properties:
                facilityId: { type: string, format: uuid }
                hicsRole: { $ref: '#/components/schemas/HicsRole' }
                isPrimaryFacility: { type: boolean, default: false }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserFacilityRole' }
  /api/users/{id}/roles/{roleId}:
    delete:
      tags: [Users]
      summary: Remove a role assignment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: roleId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Role removed }
  /api/users/{id}/sessions/{sessionId}/revoke:
    post:
      tags: [Users]
      summary: Revoke a specific session
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: sessionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Session revoked }

  # ── Facilities ────────────────────────────────────────────────────────────────
...
  /api/facilities/{id}/users:
    get:
      tags: [Facilities]
      summary: List users assigned to a facility
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 25 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array }
                  pagination: { $ref: '#/components/schemas/Pagination' }

    post:
      tags: [Facilities]
      summary: Create and add a user to this facility
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, firstName, lastName, hicsRole]
              properties:
                email: { type: string, format: email }
                firstName: { type: string }
                lastName: { type: string }
                jobTitle: { type: string }
                employeeId: { type: string }
                phoneMobile: { type: string }
                hicsRole: { $ref: '#/components/schemas/HicsRole' }
      responses:
        '201': { description: User created and invite email sent }
...